modules/location_tracker_service/requirements.txt
SQLAlchemy==1.3.19
kafka-python==2.0.2
psycopg2-binary==2.8.5
GeoAlchemy2==0.8.4
shapely==1.7.0